This is a big update. It allows main to pass around an Entity.
This was accomplished by moving away from the use of the Except monad
for error handling and instead using the ExceptT monad "transformer".
Thus, whenever we need to manage an Error, we can use the Except monad
"wrapped around" an EntityState in order to both capture an error _and_
retain the state of the Entity.

This may seem complicated, and that's because it _kind of_ is, but if
you understand monad transformers then this should make perfect sense.